+++ src/App.tsx
@@ -11,7 +11,7 @@ import StageListComponent from './components/stage/StageListComponent';
 
 import './globals.css';
 import { NavBarFunc } from './NavBarFunc';
-import VolunteerRequestComponent from './components/volunteer/VolunteerRequestComponent';
+import VolunteerFormComponent from './components/volunteer/VolunteerFormComponent';
 // import ProgramRequestSelection from './components/request/RequestProgramSelection';
 
 function App() {
@@ -35,7 +35,7 @@ function App() {
           <Route exact path="/program/list" component={ProgramListComponent} />
           <Route exact path="/stage/list" component={StageListComponent} />
           {/* <Route exact path="/beneficiaries/list" component={BeneficiariesComponent} /> */}
-          <Route exact path="/volunteer/list" component={VolunteerRequestComponent} />
+          <Route exact path="/volunteer/list" component={VolunteerFormComponent} />
           <Route component={HomeComponent} />
         </Switch>
       </div>
